moved db.js into server folder updated requires
